Web: Studio: Moved external link icon from utils to components
