+++ web/ASC.Web.Client/src/components/StyledExternalLinkIcon.js
